refactor(brain_client): rename comms -> transport; remove dead code
- Rename the comms/ folder to transport/ and update all imports, READMEs, and
  the pure-module import-rule test accordingly.
- Remove dead helpers in common/ros_services.py (await_motion_result and
  call_trigger were unused repo-wide; only call_service_sync is called) plus the
  now-unused std_srvs Trigger import.
- Drop a stale comment in nodes/skills_server.py that referenced the deleted
  ws_client node.
- Re-sort import blocks after the rename to satisfy ruff.

Verified in innate-os-test:fullbuild: colcon build, all node entry points import,
17 pure unit tests, test_fake_cloud_loop 5/5, ruff lint + format clean.
