+++ ros2_ws/src/brain/brain_client/brain_client/README.md
@@ -12,7 +12,7 @@ code, start from what it *does*:
 | `skills/` | The skill system: `registry`, `runner` (action lifecycle), `registration`, `loader`, `hot_reload`, and the public `types` SDK base classes. |
 | `agents/` | Directives/behaviours: `loader`, `initializer`, and the public `types` SDK base class. |
 | `inputs/` | Input-device subsystem and its public `types` SDK base class. |
-| `comms/` | Talking to the cloud + user: the in-process WebSocket transport (`ws_manager` + `ws_transport` + pure `ws_config`), `websocket` (the bridge with the threadsafe asyncio→executor hand-off), `messages` (typed contracts), `chat`, `tts`. |
+| `transport/` | Talking to the cloud + user: the in-process WebSocket transport (`ws_manager` + `ws_transport` + pure `ws_config`), `websocket` (the bridge with the threadsafe asyncio→executor hand-off), `messages` (typed contracts), `chat`, `tts`. |
 | `robot/` | Low-level robot facades: `mobility`, `manipulation`, `head`. |
 | `common/` | Cross-cutting leaf utilities: `logging`, `geometry`, `ros_services`, `script_paths`. |
 
@@ -21,7 +21,7 @@ code, start from what it *does*:
 **1. Dependency direction is one-way.**
 
 ```
-nodes  ->  core  ->  {perception, navigation, skills, agents, inputs, comms, robot}  ->  common
+nodes  ->  core  ->  {perception, navigation, skills, agents, inputs, transport, robot}  ->  common
 ```
 
 A module never imports "upward" (e.g. `perception` never imports `core`).
@@ -31,7 +31,7 @@ A module never imports "upward" (e.g. `perception` never imports `core`).
 Within a concept folder, intention-revealing names tell you which is which:
 `pose.py` is pure math; `pose_tracking.py` is the tf2/odom adapter. The *pure*
 files — `perception/pose`, `perception/image_codec`, `navigation/payload`,
-`skills/registry`, `comms/messages`, `core/config` — import **no `rclpy`** and are
+`skills/registry`, `transport/messages`, `core/config` — import **no `rclpy`** and are
 unit-tested without a ROS runtime. `common/import_rules_test.py` enforces this so
 the boundary can't rot.
