+++ ros2_ws/src/brain/brain_client/brain_client/common/ros_services.py
@@ -7,7 +7,6 @@ get a one-line service call instead of repeating the wait/spin/branch ladder.
 """
 
 import rclpy
-from std_srvs.srv import Trigger
 
 
 def call_service_sync(node, logger, client, request, label, timeout_sec, wait_timeout_sec=5.0):
@@ -33,39 +32,3 @@ def call_service_sync(node, logger, client, request, label, timeout_sec, wait_ti
     else:
         logger.warn(f"{label} failed: {result.message}")
     return result
-
-
-def await_motion_result(node, logger, future, label, timeout_sec):
-    """Block on an in-flight motion-service future; return True iff it succeeded."""
-    rclpy.spin_until_future_complete(node, future, timeout_sec=timeout_sec)
-    result = future.result()
-    if result is None:
-        logger.error(f"{label} call timed out")
-        return False
-    if not result.success:
-        logger.error(f"{label} returned failure")
-        return False
-    return True
-
-
-def call_trigger(node, logger, client, label, success_msg, timeout_sec=2.0):
-    """Call a std_srvs/Trigger service, log the outcome, and return whether it succeeded."""
-    if not client.service_is_ready():
-        logger.error(f"{label} service not ready")
-        return False
-
-    try:
-        future = client.call_async(Trigger.Request())
-        rclpy.spin_until_future_complete(node, future, timeout_sec=timeout_sec)
-        result = future.result()
-        if result is None:
-            logger.error(f"{label} service call timed out")
-            return False
-        if not result.success:
-            logger.error(f"{label} failed: {result.message}")
-            return False
-        logger.info(success_msg)
-        return True
-    except Exception as e:
-        logger.error(f"Exception calling {label}: {e}")
-        return False
