+++ ros2_ws/src/brain/brain_client/brain_client/skills/registration.py
@@ -13,8 +13,8 @@ import json
 from brain_messages.msg import AvailableSkills
 from rclpy.qos import QoSDurabilityPolicy, QoSProfile, QoSReliabilityPolicy
 
-from brain_client.comms.messages import MessageIn, MessageInType
 from brain_client.skills.registry import SkillRegistry
+from brain_client.transport.messages import MessageIn, MessageInType
 
 
 class SkillCatalog:
