+++ ros2_ws/src/brain/brain_client/brain_client/transport/ws_config_test.py
@@ -1,6 +1,6 @@
 """Unit tests for pure WebSocket URI/token validation (no ROS required)."""
 
-from brain_client.comms import ws_config
+from brain_client.transport import ws_config
 
 
 def test_validate_ws_uri():
